isomorphic test started

+++ bullet/bullet_interface/bullet_client_interface.go
@@ -1,6 +1,6 @@
 package bullet_interface
 
-// VX:TODO This is missing many depot bullet calls.
+// VX:TODO This is missing many depot bullet calls, but I don't think I'm going ot use bullet directly.. yet.
 type DepotClientInterface interface {
 	DepotInsertOne(req DepotRequest) error
 	DepotGetMany(req DepotGetManyRequest) (*DepotGetManyResponse, error)
@@ -14,7 +14,7 @@ type WayFinderClientInterface interface {
 
 type TrackClientInterface interface {
 	TrackGetMany(req TrackGetManyRequest) (*TrackGetManyResponse, error)
-	TrackInsertOne(bucketID int32, key string, value int, tag *int64, metric *float64) error
+	TrackInsertOne(bucketID int32, key string, value int64, tag *int64, metric *float64) error
 	TrackDeleteMany(req TrackDeleteMany) error
 	TrackGetManyByPrefix(req TrackGetItemsByPrefixRequest) (*TrackGetManyResponse, error)
 }
